2025/07/14 Progress (#3)
* Update backend URL and client configuration in UI5 setup

Modifies the backend service URL and client ID in the server configuration for improved connectivity.

Adjusts the main service URI in the manifest for proper data fetching.

These changes enhance the application's integration with the backend services.

* Adds default value for purchase order input field

Sets a default value of "4500000252" for the purchase order input field to streamline user input and improve usability.

* Replace Button with ToggleButton for status confirmation

Updates the UI component for confirming status by replacing the Button with a ToggleButton.

This change enhances user interaction by providing a more intuitive interface.

Implements an event handler for the new ToggleButton to streamline the confirmation process.

* Fixes purchase order path in MainScreen controller

Updates the purchase order path in the MainScreen controller to align with the latest API specification.

This change ensures the application retrieves the correct data for purchase orders by modifying the endpoint structure.

* Implements purchase order item selection and posting logic

Enhances the purchase order management by adding logic to handle selected order items for posting to the API.

Introduces validation to ensure items are selected and quantities are entered before proceeding.

Updates the UI interaction for confirming item status and integrates data posting functionality to improve user experience with order submissions.

* added a quantity unit column and styling the column width

* add VH for plant and storage location

* Adds file upload functionality with dialog and file handling; refactors upload and download button actions

* add refresh when posting to BAPI but still has bug

* fix value help and post to backend

* change po default value

* change entity view to ZR_RF_PO_ITEM_MAIN

* create message popover

* add ts ignore to error